Move out TextBox to its own class
